requirements.txt
aiohttp==3.8.4
aiomysql==0.1.1
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
discord.py==2.3.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
oauthlib==3.2.2
osrparse==6.0.2
ossapi==3.1.9
pycodestyle==2.10.0
PyMySQL==1.0.3
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
typing-utils==0.1.0
urllib3==2.0.2
yarl==1.9.2
